requirements.txt
Flask==2.0.2
Flask-Ngrok==0.0.25
sentence-transformers==2.1.0
spacy==3.1.3
newspaper3k